add settings example
