+++ experimental/schema/query.go
@@ -1,4 +1,4 @@
-package query
+package schema
 
 import (
 	"embed"
@@ -6,6 +6,50 @@ import (
 
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 )
+
+type QueryTypeDefinitionSpec struct {
+	// DiscriminatorField is the field used to link behavior to this specific
+	// query type.  It is typically "queryType", but can be another field if necessary
+	DiscriminatorField string `json:"discriminatorField,omitempty"`
+
+	// The discriminator value
+	DiscriminatorValue string `json:"discriminatorValue,omitempty"`
+
+	// Describe whe the query type is for
+	Description string `json:"description,omitempty"`
+
+	// The query schema represents the properties that can be sent to the API
+	// In many cases, this may be the same properties that are saved in a dashboard
+	// In the case where the save model is different, we must also specify a save model
+	QuerySchema any `json:"querySchema"`
+
+	// The save model defines properties that can be saved into dashboard or similar
+	// These values are processed by frontend components and then sent to the api
+	// When specified, this schema will be used to validate saved objects rather than
+	// the query schema
+	SaveModel any `json:"saveModel,omitempty"`
+
+	// Examples (include a wrapper) ideally a template!
+	Examples []QueryExample `json:"examples,omitempty"`
+
+	// Changelog defines the changed from the previous version
+	// All changes in the same version *must* be backwards compatible
+	// Only notable changes will be shown here, for the full version history see git!
+	Changelog []string `json:"changelog,omitempty"`
+}
+
+type QueryExample struct {
+	// Version identifier or empty if only one exists
+	Name string `json:"name,omitempty"`
+
+	// An example payload -- this should not require the frontend code to
+	// pre-process anything
+	QueryPayload any `json:"queryPayload,omitempty"`
+
+	// An example save model -- this will require frontend code to convert it
+	// into a valid query payload
+	SaveModel any `json:"saveModel,omitempty"`
+}
 
 type CommonQueryProperties struct {
 	// RefID is the unique identifier of the query, set by the frontend call.
@@ -16,6 +60,7 @@ type CommonQueryProperties struct {
 
 	// TimeRange represents the query range
 	// NOTE: unlike generic /ds/query, we can now send explicit time values in each query
+	// NOTE: the values for timeRange are not saved in a dashboard, they are constructed on the fly
 	TimeRange *TimeRange `json:"timeRange,omitempty"`
 
 	// The datasource
@@ -29,13 +74,17 @@ type CommonQueryProperties struct {
 	QueryType string `json:"queryType,omitempty"`
 
 	// MaxDataPoints is the maximum number of data points that should be returned from a time series query.
+	// NOTE: the values for maxDataPoints is not saved in the query model.  It is typically calculated
+	// from the number of pixels visible in a visualization
 	MaxDataPoints int64 `json:"maxDataPoints,omitempty"`
 
 	// Interval is the suggested duration between time points in a time series query.
+	// NOTE: the values for intervalMs is not saved in the query model.  It is typically calculated
+	// from the interval required to fill a pixels in the visualization
 	IntervalMS float64 `json:"intervalMs,omitempty"`
 
 	// true if query is disabled (ie should not be returned to the dashboard)
-	// Note this does not always imply that the query should not be executed since
+	// NOTE: this does not always imply that the query should not be executed since
 	// the results from a hidden query may be used as the input to other queries (SSE etc)
 	Hide bool `json:"hide,omitempty"`
 }
@@ -78,15 +127,7 @@ type ResultAssertions struct {
 	MaxFrames int64 `json:"maxFrames,omitempty"`
 }
 
-// GenericDataQuery is a replacement for `dtos.MetricRequest` with more explicit typing
-type GenericDataQuery struct {
-	CommonQueryProperties `json:",inline"`
-
-	// Additional Properties (that live at the root)
-	Additional map[string]any `json:",inline"`
-}
-
-//go:embed common.jsonschema
+//go:embed query.schema.json
 var f embed.FS
 
 // Get the cached feature list (exposed as a k8s resource)
